+++ src/main/java/com/mddb/usecase/impl/DeviceServiceMock.java
@@ -35,7 +35,7 @@ public class DeviceServiceMock implements DeviceService {
     }
 
     @Override
-    public Map<Long, String> getDevicesByCompany(String company) {
+    public Map<Integer, String> getDevicesByCompany(String company) {
         return repository.findByCompanyName(company);
     }
